GBB.ConversationalKM.Python/requirements.txt
azure-functions
azure-cosmosdb-table
pandas
azure-storage-blob==12.3.0
requests